backend/requirements.txt
fastapi
uvicorn
python-dotenv
mindsdb_sdk
pydantic
